pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "mylene", email = "mylene@gmail.com"},
]
dependencies = [
    "django>=4.2.1",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
